requirements.txt
Flask==3.0.3
numpy==1.26.4
pillow==11.0.0
h5py==3.12.1
tensorflow-cpu==2.17.0
